requirements.txt
python-dotenv==1.0.1
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
openai==0.28.0
tiktoken==0.9.0
sentence_transformers===3.4.1
tokenizers===0.21.0
faiss-cpu===1.10.0
unstructured===0.16.23
numpy===1.26.4
nltk===3.9.1
-e .
fastapi==0.115.8
uvicorn==0.34.0
jinja2==3.1.5
python-multipart==0.0.20
